results/rebuttal_new_baselines/run-20231112-040610_CodeT-SelfRefine-Reflexion-ZeroShot_donnemartin-system-design-oop-recipe-donnemartin-system-design-oop-eventplanner_0_20-runs_log/recipe/9/Reflexion/test_app.py
import pytest
import app

@pytest.fixture
def client():
	app.app.config['TESTING'] = True
	with app.app.test_client() as client:
		yield client


def test_create_user(client):
	response = client.post('/user', json={'name': 'John Doe', 'email': 'john@example.com'})
	assert response.status_code == 201
	assert response.get_json() == {'message': 'User created successfully'}


def test_create_recipe(client):
	client.post('/user', json={'name': 'John Doe', 'email': 'john@example.com'})
	response = client.post('/recipe', json={'title': 'Pasta', 'instructions': 'Boil water. Cook pasta.', 'ingredients': 'Pasta, Water', 'user_email': 'john@example.com'})
	assert response.status_code == 201
	assert response.get_json() == {'message': 'Recipe created successfully'}


def test_create_review(client):
	client.post('/user', json={'name': 'John Doe', 'email': 'john@example.com'})
	client.post('/recipe', json={'title': 'Pasta', 'instructions': 'Boil water. Cook pasta.', 'ingredients': 'Pasta, Water', 'user_email': 'john@example.com'})
	response = client.post('/review', json={'user_email': 'john@example.com', 'recipe_title': 'Pasta', 'comment': 'Delicious!'})
	assert response.status_code == 201
	assert response.get_json() == {'message': 'Review created successfully'}
